backend/services/identity/requirements.txt
# =============================================================================
# Identity Service Requirements (Python 3.14 compatible)
# =============================================================================

# --- Web Framework -----------------------------------------------------------
fastapi>=0.115.5
uvicorn[standard]>=0.32.0

# --- Database ----------------------------------------------------------------
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.29.0
alembic>=1.14.0

# --- Security ----------------------------------------------------------------
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0

# --- Validation & Config -----------------------------------------------------
pydantic[email]>=2.11.0          # 2.11+ has Python 3.14 pre-built wheels
python-dotenv>=1.0.1

# --- Shared package (installed separately) -----------------------------------
# pip install -e ../../../shared
